+++ Form/Type/GroupFormType.php
@@ -11,10 +11,15 @@
 
 namespace FOS\UserBundle\Form\Type;
 
+@trigger_error('Using Groups is deprecated since version 2.2 and will be removed in 3.0.', E_USER_DEPRECATED);
+
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
+/**
+ * @deprecated
+ */
 class GroupFormType extends AbstractType
 {
     /**