+++ Model/GroupInterface.php
@@ -11,9 +11,13 @@
 
 namespace FOS\UserBundle\Model;
 
+@trigger_error('Using Groups is deprecated since version 2.2 and will be removed in 3.0.', E_USER_DEPRECATED);
+
 /**
  * @author Johannes M. Schmitt <schmittjoh@gmail.com>
  * @author Christophe Coevoet <stof@notk.org>
+ *
+ * @deprecated
  */
 interface GroupInterface
 {